% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratPlot_varlist.R
\name{stratPlot_varlist}
\alias{stratPlot_varlist}
\title{stratPlot_varlist}
\usage{
stratPlot_varlist(var_list)
}
\arguments{
\item{var_list, }{a list with the original data split into n list items (variable-wise)}
}
\value{
gridded stratigraphic plot of with all the variables and height of top strip that has group labels
}
\description{
makes stratigraphic plots for the n variables in the input data
}

